# docs/swagger.yaml
basePath: /
definitions:
  handler.FieldValidationError:
    properties:
      error:
        type: string
      name:
        type: string
    type: object
  handler.HTTPError:
    properties:
      details:
        type: string
    type: object
  handler.ValidationError:
    properties:
      fields:
        items:
          $ref: '#/definitions/handler.FieldValidationError'
        type: array
    type: object
  model.AuthCredentials:
    properties:
      code:
        example: "6666"
        type: string
      user_id:
        example: johndoe@example.com
        type: string
    type: object
  model.CodeRequest:
    properties:
      email:
        example: johndoe@example.com
        type: string
    type: object
  model.InviteCreate:
    properties:
      from_user:
        example: 1
        type: integer
      to_organization:
        example: 1
        type: integer
      user_email:
        example: johndoe@example.com
        type: string
    type: object
  model.InviteGet:
    properties:
      from_user:
        example: 1
        type: integer
      invite_id:
        example: 1
        type: integer
      status:
        enum:
        - sent
        - accepted
        - rejected
        type: string
      to_organization:
        example: 1
        type: integer
      user_email:
        example: johndoe@example.com
        type: string
      user_id:
        example: 2
        type: integer
    type: object
  model.OrganizationCreate:
    properties:
      address:
        example: Г. Москва, Пр-т. Вернадского 78
        type: string
      contact_email:
        example: contact@mirea.ru
        type: string
      contact_phone:
        example: "74992156565"
        type: string
      name:
        example: Российский технологический университет МИРЭА
        type: string
    type: object
  model.OrganizationGet:
    properties:
      address:
        example: Г. Москва, Пр-т. Вернадского 78
        type: string
      contact_email:
        example: contact@mirea.ru
        type: string
      contact_phone:
        example: "74992156565"
        type: string
      name:
        example: Российский технологический университет МИРЭА
        type: string
      organization_id:
        example: 1
        type: integer
    type: object
  model.OrganizationUpdate:
    properties:
      address:
        example: Г. Москва, Пр-т. Вернадского 78
        type: string
      contact_email:
        example: contact@mirea.ru
        type: string
      contact_phone:
        example: "74992156565"
        type: string
      name:
        example: Российский технологический университет МИРЭА
        type: string
    type: object
  model.UserCreate:
    properties:
      email:
        example: johndoe@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      middle_name:
        example: Jr.
        type: string
    type: object
  model.UserGet:
    properties:
      email:
        example: johndoe@example.com
        type: string
      first_name:
        example: John
        type: string
      is_active:
        example: true
        type: boolean
      last_name:
        example: Doe
        type: string
      middle_name:
        example: Jr.
        type: string
      user_id:
        example: 1
        type: integer
    type: object
host: localhost:8000
info:
  contact:
    email: burenotti@gmail.com
    name: Буренин Артём
  description: Реализация тестового задания для RTUITLab.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: API системы управления городскими меропреятиями
  version: 0.1.0
paths:
  /auth/activate/{token}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Activation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Activates user with token sent in email
      tags:
      - Auth
  /auth/request:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CodeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Requests sending one time password to users email
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.AuthCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Signs user in using sent in email one time password
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserGet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Creates new user that should be activated with email
      tags:
      - Auth
  /organization/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.OrganizationCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.OrganizationGet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Creates a new organization
      tags:
      - Organizations
  /organization/{organization_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Organization id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Deletes organization by id
      tags:
      - Organizations
    get:
      consumes:
      - application/json
      parameters:
      - description: Organization id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrganizationGet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Returns an information about organization
      tags:
      - Organizations
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization id
        in: path
        name: organization_id
        required: true
        type: integer
      - description: Fields that will be updated
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/model.OrganizationUpdate'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ValidationError'
      security:
      - APIKey: []
      summary: Updates organization information
      tags:
      - Organizations
  /organization/{organization_id}/invite/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/model.InviteGet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Returns list of invites to organization
      tags:
      - Invites
  /organization/{organization_id}/invite/{invite_id}/accept:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite id
        in: path
        name: invite_id
        required: true
        type: integer
      - description: Organization id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Current user accepts invite and joins organization
      tags:
      - Invites
  /organization/{organization_id}/invite/{invite_id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite id
        in: path
        name: invite_id
        required: true
        type: integer
      - description: Organization id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Current user rejects invite and joins organization
      tags:
      - Invites
  /organization/{organization_id}/leave:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Invite id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Current user leaves organization
      tags:
      - Members
  /organization/{organization_id}/member/{member_id}/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Member id
        in: path
        name: member_id
        required: true
        type: integer
      - description: Organization id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Removes member from organization
      tags:
      - Members
    put:
      consumes:
      - application/json
      parameters:
      - description: Member id
        in: path
        name: member_id
        required: true
        type: integer
      - description: Organization id
        in: path
        name: organization_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - APIKey: []
      summary: Updates organization member's privileges
      tags:
      - Members
securityDefinitions:
  APIKey:
    description: OAuth protects our entity endpoints
    in: header
    name: APIKey
    type: apiKey
swagger: "2.0"